+++ skmq/postoffice.go
@@ -1,4 +1,4 @@
-package session
+package skmq
 
 import (
 	"net"
